+++ JSM.Insight/Public/Set-InsightObject.ps1
@@ -21,8 +21,10 @@ function Set-InsightObject {
 
         $RequestBody = @{
             'objectTypeId' = $objectTypeId
-            'attributes'   = @($attributes)
+            'attributes'   = @($attributesArray)
             }
+
+        $RequestBody = $RequestBody | ConvertTo-Json -Depth 5
 
         $Request = [System.UriBuilder]"https://api.atlassian.com/jsm/insight/workspace/$InsightWorkspaceID/v$Version/object/$id"
     }